[refactor] Added a top-level `jscoq.js`.
To avoid confusion, `ui-js/jscoq.js` has been renamed.
Also, a legacy-mode `jscoq-loader.js` that emulates the old behavior.
