+++ docs/npm-landing.html
@@ -26,7 +26,7 @@ Inductive and_they_run := too(*!*).
   <!-- jsCoq configuration part -->
   <script type="module">
 
-    import { JsCoq } from './ui-js/index.js';
+    import { JsCoq } from './jscoq.js';
 
     var jscoq_ids  = ['.coq-code'];
     var jscoq_opts = {
