+++ examples/scratchpad.html
@@ -21,7 +21,7 @@
   </div>
 
   <script type="module">
-    import { JsCoq } from "../ui-js/index.js";
+    import { JsCoq } from "../jscoq.js";
 
     var sp = new URLSearchParams(location.search),
         panel_theme = sp.has('dark') ? 'dark' : sp.get('theme') || 'light',
