+++ index.html
@@ -247,7 +247,7 @@ Qed.</textarea>
 
   <script type="module">
 
-    import { JsCoq } from './ui-js/index.js';
+    import { JsCoq } from './jscoq.js';
 
     if (!localStorage['scratchpad.last_filename'])
       setTimeout(() => document.body.classList.add('welcome'), 1500);
