+++ ui-js/coq-manager.js
@@ -13,7 +13,7 @@ import { $ } from '../dist/lib.js';
 
 import { JsCoq } from './index.js';
 import { copyOptions } from './etc.js';
-import { Future, CoqWorker } from './jscoq.js';
+import { Future, CoqWorker } from './jscoq-worker-interface.js';
 import { PackageManager } from './coq-packages.js';
 import { CoqLayoutClassic } from './coq-layout-classic.js';
 import { ProviderContainer } from './cm-provider-container.js';
